requirements.txt
flask
openai
python-dotenv
requests
PyMuPDF
